move util dependency install to helpers feature

+++ features/src/dcc-helpers/cont.sh
@@ -14,17 +14,22 @@ fi
 
 echo "continuously running '$command' watching '$watch'"
 
-inotifywait -q -m -e close_write,create --recursive $watch | \
-  (
-    while
-        if [ ! -z "$PID" ] && ps -p $PID > /dev/null
-        then
-            kill -9 $PID
-        fi
-        $command <&1 &
-        PID=$!
-        read -r filename event
-    do
-        :
-    done
-  )
+if command -v onchange >/dev/null
+then
+    onchange -ik "$watch" -- "$command"
+else
+    inotifywait -q -m -e close_write,create --recursive $watch | \
+    (
+        while
+            if [ ! -z "$PID" ] && ps -p $PID > /dev/null
+            then
+                kill -9 $PID
+            fi
+            $command <&1 &
+            PID=$!
+            read -r filename event
+        do
+            :
+        done
+    )
+fi
